refactor pages to be module inside ui instead of components
