+++ src/ui/mod.rs
@@ -3,6 +3,7 @@ pub mod assets;
 pub mod components;
 pub mod global_keybinds;
 pub mod helpers;
+pub mod pages;
 pub mod res_handler;
 pub mod theme;
 pub mod wiremann;
