+++ src/ui/pages/mod.rs
@@ -1,5 +1,3 @@
 pub mod library;
-pub mod library_helpers;
 pub mod player;
 pub mod playlists;
-pub mod playlists_helpers;
